Move permission check to core interface user
