+++ Scripts/Command/Command.cs
@@ -1,15 +1,16 @@
 ﻿namespace Stonebot.Scripts.Command {
     using Bot_Core.Models.EventSub;
+    using Core_Interface;
     using System;
     using System.Text.Json;
     using System.Threading.Tasks;
 
     internal class Command {
-        public event EventHandler<PermissionLevel> PermissionLevelChanged = delegate { };
+        public event EventHandler<User.PermissionLevel> PermissionLevelChanged = delegate { };
         public event EventHandler<int> UseDelayChanged = delegate { };
 
         public string Keyword { get; private set; }
-        public PermissionLevel PermissionLevel { get => permissionLevel; set => SetPermissionLevel(value); }
+        public User.PermissionLevel PermissionLevel { get => permissionLevel; set => SetPermissionLevel(value); }
         public int UseDelay { get => useDelay; set => SetUseDelay(value); }
         public DateTime LastUsed { get; private set; } = DateTime.Now;
         public Func<ChannelChatMessageEvent, Task<bool>> UseAction;
@@ -31,9 +32,9 @@
                 return false;
             }
 
-            var userPermissionLevel = await Permission.GetHighest(messageEvent.ChatterUserId);
+            var userPermissionLevel = await User.GetHighestPermissionLevel(messageEvent.ChatterUserId);
             if (userPermissionLevel is null) {
-                Logger.Warning($"{logPrefix} | {nameof(Permission.GetHighest)} result is null.");
+                Logger.Warning($"{logPrefix} | {nameof(User.GetHighestPermissionLevel)} result is null.");
                 return null;
             }
 
@@ -50,7 +51,7 @@
             return true;
         }
 
-        public void SetPermissionLevel(PermissionLevel permissionLevel) {
+        public void SetPermissionLevel(User.PermissionLevel permissionLevel) {
             Logger.Info($"{nameof(Command)} | {nameof(SetPermissionLevel)}\n{nameof(permissionLevel)}: {permissionLevel}");
 
             this.permissionLevel = permissionLevel;
@@ -64,7 +65,7 @@
             Util.InvokeDeferred(UseDelayChanged, UseDelay);
         }
 
-        private PermissionLevel permissionLevel = PermissionLevel.Viewer;
+        private User.PermissionLevel permissionLevel = User.PermissionLevel.Viewer;
         private int useDelay = 1000;
     }
